refactor(v2): extract action offer composition
